#49 - extracting reserve item to another file
