+++ feature/misc/src/main/java/br/com/rbrthmn/misc/reserves/ReservesScreen.kt
@@ -2,9 +2,7 @@ package br.com.rbrthmn.misc.reserves
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.animation.core.LinearOutSlowInEasing
-import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
@@ -18,7 +16,6 @@ import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -29,13 +26,8 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
@@ -48,15 +40,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.misc.R
+import br.com.rbrthmn.misc.reserves.components.ReserveItem
 import br.com.rbrthmn.navigation.NavigationDestination
-import br.com.rbrthmn.operations.ui.OperationType
-import br.com.rbrthmn.operations.ui.OperationsScreenContract
-import br.com.rbrthmn.operations.ui.components.AddOperationDialog
 import br.com.rbrthmn.ui.utils.valueWithCurrencyString
 import org.koin.androidx.compose.koinViewModel
-import java.text.SimpleDateFormat
-import java.util.Locale
-import br.com.rbrthmn.operations.R as operationR
 import br.com.rbrthmn.ui.R as commonR
 
 object ReservesDestination : NavigationDestination {
@@ -279,172 +266,6 @@ private fun NewReserveDialog(
                         Text(text = stringResource(id = commonR.string.save_button))
                     }
                 }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ReserveItem(
-    modifier: Modifier = Modifier,
-    reserve: Reserve,
-    isExpanded: Boolean,
-    onClick: () -> Unit
-) {
-    val rotationState by animateFloatAsState(
-        targetValue = if (isExpanded) 180f else 0f, label = ""
-    )
-
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(dimensionResource(id = commonR.dimen.corner_shape_round)))
-            .clickable { onClick() }
-            .padding(end = dimensionResource(id = commonR.dimen.padding_extra_small)),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically, modifier = modifier.weight(1f)) {
-            Icon(
-                imageVector = Icons.Default.ArrowDropDown,
-                contentDescription = stringResource(id = commonR.string.drop_down_arrow_icon_description),
-                modifier = modifier
-                    .alpha(0.2f)
-                    .rotate(rotationState)
-                    .padding(dimensionResource(id = commonR.dimen.padding_extra_small)),
-            )
-            Text(
-                text = reserve.name,
-                fontSize = dimensionResource(id = commonR.dimen.font_size_medium).value.sp,
-                fontWeight = FontWeight.Bold,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-        Text(
-            text = valueWithCurrencyString(
-                currencyStringId = commonR.string.brl_currency, value = reserve.value
-            ),
-            maxLines = 1,
-            fontSize = dimensionResource(id = commonR.dimen.font_size_medium).value.sp,
-            fontWeight = FontWeight.Bold,
-        )
-    }
-
-    if (isExpanded) {
-        ReserveOperationsList(reserve, modifier)
-    }
-}
-
-@Composable
-private fun ReserveOperationsList(
-    reserve: Reserve,
-    modifier: Modifier = Modifier
-) {
-    val showDialog = remember { mutableStateOf(false) }
-    val operationsScreenViewModel: OperationsScreenContract.ViewModel =
-        koinViewModel()
-
-    if (showDialog.value) {
-        AddOperationDialog(
-            uiState = operationsScreenViewModel.uiState.collectAsState().value,
-            onSaveButtonClick = {
-                operationsScreenViewModel.onIntent(
-                    OperationsScreenContract.Intent.OnSaveButtonClick(
-                        showDialog
-                    )
-                )
-            },
-            onCancelButtonClick = {
-                showDialog.value = false
-                operationsScreenViewModel.onIntent(OperationsScreenContract.Intent.OnResetDialogFields)
-            },
-            onDescriptionChange = {
-                operationsScreenViewModel.onIntent(
-                    OperationsScreenContract.Intent.OnDescriptionChange(
-                        it
-                    )
-                )
-            },
-            onTypeChange = {
-                operationsScreenViewModel.onIntent(
-                    OperationsScreenContract.Intent.OnOperationTypeChange(
-                        it
-                    )
-                )
-            },
-            onValueChange = {
-                operationsScreenViewModel.onIntent(OperationsScreenContract.Intent.OnValueChange(it))
-            },
-            onDateChange = {
-                operationsScreenViewModel.onIntent(
-                    OperationsScreenContract.Intent.OnOperationDateChange(
-                        it
-                    )
-                )
-            },
-            availableOperationTypes = listOf(
-                OperationType.RESERVE_ALLOCATION,
-                OperationType.RESERVE_WITHDRAWAL
-            )
-        )
-    }
-
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(dimensionResource(id = commonR.dimen.padding_extra_small))
-    ) {
-        for (operation in reserve.operations) {
-            val operationValueColor = if (operation.isWithdrawal) Color.Red else Color.Green
-            val operationValueSymbol = if (operation.isWithdrawal) "-" else "+"
-
-            Row(
-                modifier = modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = dimensionResource(id = commonR.dimen.padding_medium))
-                    .padding(vertical = dimensionResource(id = commonR.dimen.padding_extra_small)),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.SpaceBetween
-            ) {
-                Text(
-                    text = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault()).format(
-                        SimpleDateFormat(
-                            "yyyy-MM-dd",
-                            Locale.getDefault()
-                        ).parse(operation.date)!!
-                    ),
-                    fontSize = dimensionResource(id = commonR.dimen.font_size_medium).value.sp
-                )
-                Text(
-                    text = operationValueSymbol + valueWithCurrencyString(
-                        currencyStringId = commonR.string.brl_currency,
-                        value = operation.value
-                    ),
-                    color = operationValueColor,
-                    fontSize = dimensionResource(id = commonR.dimen.font_size_medium).value.sp
-                )
-            }
-        }
-        TextButton(
-            onClick = { showDialog.value = true },
-            modifier = modifier
-                .fillMaxWidth()
-                .padding(horizontal = dimensionResource(id = commonR.dimen.padding_extra_small))
-        ) {
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.Start,
-                modifier = modifier.fillMaxWidth()
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = stringResource(id = commonR.string.add_icon_description),
-                    tint = Color.Gray,
-                )
-                Text(
-                    text = stringResource(id = operationR.string.add_operation_button),
-                    fontSize = dimensionResource(id = commonR.dimen.font_size_medium).value.sp
-                )
             }
         }
     }
